security question for deleting posts, styling changed

+++ src/components/PostSettingModal/PostSettingModal.tsx
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { mainContext } from "../../context/MainProvider";
 import supabase from "../../utils/supabase";
 import IPost from "../../interfaces/IPost";
@@ -20,6 +20,8 @@ const PostSettingModal = ({
     setShowPostSettingModal(false);
   };
 
+  const[security, setSecurity] = useState<boolean>(false)
+
   const handleEdit = () => {
     // Speichere den Post im Session Storage für die Bearbeitung
     sessionStorage.setItem("editPost", JSON.stringify(post));
@@ -31,24 +33,33 @@ const PostSettingModal = ({
 
   const handleDelete = async () => {
     // Lösche den Post aus der Datenbank
-    await supabase.from("posts").delete().eq("id", post.id);
-    // Optional: Seite neu laden oder Post aus dem State entfernen
-    setShowPostSettingModal(false);
-    window.location.reload();
+    if (security === false) {
+        setSecurity(true)
+    } else {
+        await supabase.from("posts").delete().eq("id", post.id);
+      // Optional: Seite neu laden oder Post aus dem State entfernen
+        setShowPostSettingModal(false);
+        setSecurity(false)
+        window.location.reload();
+        console.log(security)
+    }
   };
 
+
   return (
-    <section
-      className="inset-0 fixed bg-white/50 shadow-2xl"
-      onClick={closeModule}
-    >
-      <div className=" flex flex-col gap-8 items-center justify-start p-5 rounded-4xl absolute w-1/2 h-1/6 top-40 left-[25%] bg-white/90 ">
+    <>
+    <div
+    onClick={closeModule}
+    className="inset-0 z-10 fixed bg-white/50">
+    </div>
+
+    <div className="z-20 flex flex-col gap-5 items-center justify-start p-5 rounded-4xl absolute w-1/2 h-1/5 top-40 left-[25%] bg-white/90 ">
         <h1 className="text-xl font-bold">Post Settings</h1>
         {user && user.id === post.user_id && (
           <div className="flex items-center justify-around gap-5">
             <div>
               <button
-                className=" text-xl cursor-pointer transition ease-in-out hover:font-bold"
+                className=" text-xl p-5 cursor-pointer transition ease-in-out hover:font-bold"
                 onClick={handleEdit}
               >
                 edit
@@ -57,16 +68,18 @@ const PostSettingModal = ({
 
             <div>
               <button
-                className=" text-xl cursor-pointer transition ease-in-out hover:font-bold hover:text-red-500"
+                className="p-5 text-xl cursor-pointer transition ease-in-out hover:font-bold hover:text-red-500"
                 onClick={handleDelete}
               >
-                delete
+                {security 
+                ? "really delete?"
+                : "delete"}
               </button>
             </div>
           </div>
         )}
       </div>
-    </section>
+    </>
   );
 };
 export default PostSettingModal;
